requirements.txt
importlib-metadata==4.12.0
Markdown==3.4.1
python-dotenv==0.20.0
zipp==3.8.1
